pyproject.toml
[project]
name = "ravelights"
version = "0.1.0"
requires-python = "~=3.10"
dependencies = [
    "numpy == 1.24.1",
    "plum-py == 0.8.5",
    "Flask == 2.2.2",
    "Flask-RESTful == 0.3.9",
]

[project.optional-dependencies]
dev = [
    "pytest==7.2.1",
    "black==22.3.0",
    "flake8==4.0.1",
    "mypy==0.991",
]
gui = [
    "pygame == 2.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*"]

[build-system]
requires = ["setuptools>=42.0", "wheel", "pip >= 21.1"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 120
